requirements.txt
apache_beam
google-cloud-secret-manager
google-cloud-bigquery
teradatasql
apache_beam[dataframe]
